refactor(goesdr): add Imager Projection module
Move GOES Imager Projection classes from goesdr module to projection
module.
